common: refactor jwtUtils, frontend: Material table, ConfirmDialog, resource-reservation: filter reservation for users/admins

+++ resource-reservation/src/main/java/com/resourcify/service/ResourceService.java
@@ -17,40 +17,42 @@ import java.util.List;
 @Transactional
 public class ResourceService {
 
-  public final ResourceRepository resourceRepository;
-  private final ResourceMapper resourceMapper;
+    public final ResourceRepository resourceRepository;
+    private final ResourceMapper resourceMapper;
 
-  // TODO: remove !isActive reservations
-  public List<ResourceResponse> findAll() {
-    List<Resource> resources = resourceRepository.findAll();
-    return resources.stream().map(resourceMapper::toResponse).toList();
-  }
+    // TODO: remove !isActive reservations
+    public List<ResourceResponse> findAll(boolean isAdmin) {
+        List<Resource> resources = resourceRepository.findAll();
+        return resources.stream().map(resource -> resourceMapper.toResponse(resource, isAdmin)).toList();
+    }
 
-  public ResourceResponse findById(String id) {
-    Resource resource = resourceRepository.findById(id)
-        .orElseThrow(() -> new NotFoundException(Resource.class, id));
-    return resourceMapper.toResponse(resource);
-  }
+    public ResourceResponse findById(String id, boolean isAdmin) {
+        Resource resource = resourceRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException(Resource.class, id));
+        return resourceMapper.toResponse(resource, isAdmin);
+    }
 
-  public ResourceResponse insert(ResourceRequest resourceRequest) {
-    Resource resource = resourceMapper.fromRequest(resourceRequest);
-    resource = resourceRepository.save(resource);
-    return resourceMapper.toResponse(resource);
-  }
+    public ResourceResponse insert(ResourceRequest resourceRequest) {
+        Resource resource = resourceMapper.fromRequest(resourceRequest);
+        resource = resourceRepository.save(resource);
+        return resourceMapper.toResponse(resource);
+    }
 
-  public ResourceResponse update(String id, ResourceRequest resourceRequest) {
-    Resource resource = resourceRepository.findById(id)
-        .orElseThrow(() -> new NotFoundException(Resource.class, id));
-    resource.setName(resourceRequest.getName());
-    resource.setDescription(resourceRequest.getDescription());
-    resource.setAmount(resourceRequest.getAmount());
-    resource.setBackgroundColor(resourceRequest.getBackgroundColor());
-    resource = resourceRepository.save(resource);
-    return resourceMapper.toResponse(resource);
-  }
+    public ResourceResponse update(String id, ResourceRequest resourceRequest) {
+        Resource resource = resourceRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException(Resource.class, id));
+        resource.setName(resourceRequest.getName());
+        resource.setDescription(resourceRequest.getDescription());
+        resource.setAmount(resourceRequest.getAmount());
+        resource.setBackgroundColor(resourceRequest.getBackgroundColor());
+        resource = resourceRepository.save(resource);
+        return resourceMapper.toResponse(resource);
+    }
 
-  public void delete(String id) {
-    findById(id);
-    resourceRepository.deleteById(id);
-  }
+    public void delete(String id) {
+        if (!resourceRepository.existsById(id)) {
+            throw new NotFoundException(Resource.class, id);
+        }
+        resourceRepository.deleteById(id);
+    }
 }
